requirements.txt
pyrebase
python_dotenv